feat: M2-767 Sing Up - change connection due to the new API update (firstName, lastName fields instead of the fullName), correct API mapping in the Redux Store
