+++ src/features/Builder/ActivityItems/ItemConfiguration/GroupedSelectSearchController/ItemTypeTooltip/TooltipComponents/index.ts
@@ -1,9 +1,9 @@
 export * from './TooltipComponents.types';
-export * from './Components/Selection';
-export * from './Components/Slider';
-export * from './Components/Date';
-export * from './Components/NumberSelection';
-export * from './Components/TimeRange';
-export * from './Components/SliderRows';
-export * from './Components/Dashed';
-export * from './Components/SelectionPerRow';
+export * from './Selection';
+export * from './Slider';
+export * from './Date';
+export * from './NumberSelection';
+export * from './TimeRange';
+export * from './SliderRows';
+export * from './Dashed';
+export * from './SelectionPerRow';
